Fix warnings.
Since bumping Qt Creator version to 4.70 many warnings appeared in the
IDE.

- Implicit conversions between float and double.
- No out-of-line virtual method definitions.
- Shadowed class variables (mainly overriding Character with <Class>).
- Old-style casting between data types.
- Using 0 instead of nullptr for QObject pointers in constructors
(this is how Qt used to document constructors).

Also:
- Remove AutoAttack.h; it is unused.
- Fixing implicit conversions seem to have been enough to fix Execute
tests which previously had off-by-one errors when asserting damage.
- However, Deep Wounds tests now have off-by-one errors instead.

+++ Character/CharacterStats.cpp
@@ -51,13 +51,13 @@ int CharacterStats::get_spirit(void) const {
     return base_stats->get_spirit() + equipment->get_stats()->get_spirit() + pchar->get_race()->get_base_spirit();
 }
 
-float CharacterStats::get_hit_chance(void) const {
+double CharacterStats::get_hit_chance(void) const {
     return base_stats->get_hit_chance()  + equipment->get_stats()->get_hit_chance();
 }
 
-float CharacterStats::get_crit_chance(void) const {
-    const float equip_effect = base_stats->get_crit_chance()  + equipment->get_stats()->get_crit_chance();
-    const float crit_from_agi = float(float(get_agility()) / pchar->get_agi_needed_for_one_percent_phys_crit());
+double CharacterStats::get_crit_chance(void) const {
+    const double equip_effect = base_stats->get_crit_chance()  + equipment->get_stats()->get_crit_chance();
+    const double crit_from_agi = double(double(get_agility()) / pchar->get_agi_needed_for_one_percent_phys_crit());
 
     return equip_effect + crit_from_agi / 100;
 }
@@ -67,7 +67,7 @@ void CharacterStats::increase_haste(const int increase) {
 
     haste_factor = 1.0;
     for (int i = 0; i < attack_speed_buffs.size(); ++i) {
-        float haste_buff = float(attack_speed_buffs[i]) / 100;
+        double haste_buff = double(attack_speed_buffs[i]) / 100;
         haste_factor *= 1 + haste_buff;
     }
 }
@@ -77,7 +77,7 @@ void CharacterStats::decrease_haste(const int decrease) {
 
     haste_factor = 1.0;
     for (int i = 0; i < attack_speed_buffs.size(); ++i) {
-        float haste_buff = float(attack_speed_buffs[i]) / 100;
+        double haste_buff = double(attack_speed_buffs[i]) / 100;
         haste_factor *= 1 + haste_buff;
     }
 }
@@ -140,27 +140,27 @@ void CharacterStats::decrease_melee_ap(const int decrease) {
     base_stats->decrease_base_melee_ap(decrease);
 }
 
-float CharacterStats::get_total_phys_dmg_mod() const {
+double CharacterStats::get_total_phys_dmg_mod() const {
     // TODO: Include e.g. Two-Handed Weapon Specialization if using 2-handers.
     return total_phys_dmg_mod;
 }
 
-void CharacterStats::increase_hit(float increase) {
+void CharacterStats::increase_hit(double increase) {
     base_stats->increase_hit(increase);
     pchar->get_combat_roll()->update_miss_chance(get_hit_chance());
 }
 
-void CharacterStats::decrease_hit(float decrease) {
+void CharacterStats::decrease_hit(double decrease) {
     base_stats->decrease_hit(decrease);
     pchar->get_combat_roll()->update_miss_chance(get_hit_chance());
 }
 
-void CharacterStats::increase_crit(float increase) {
+void CharacterStats::increase_crit(double increase) {
     base_stats->increase_crit(increase);
     pchar->get_combat_roll()->update_crit_chance(get_crit_chance());
 }
 
-void CharacterStats::decrease_crit(float decrease) {
+void CharacterStats::decrease_crit(double decrease) {
     base_stats->decrease_crit(decrease);
     pchar->get_combat_roll()->update_crit_chance(get_crit_chance());
 }
@@ -170,7 +170,7 @@ void CharacterStats::increase_total_phys_dmg_mod(const int increase) {
 
     total_phys_dmg_mod = 1.0;
     for (int i = 0; i < phys_dmg_buffs.size(); ++i) {
-        float coefficient = 1.0 + float(phys_dmg_buffs[i]) / 100;
+        double coefficient = 1.0 + double(phys_dmg_buffs[i]) / 100;
         total_phys_dmg_mod = total_phys_dmg_mod * coefficient;
     }
 }
@@ -180,17 +180,17 @@ void CharacterStats::decrease_total_phys_dmg_mod(const int decrease) {
 
     total_phys_dmg_mod = 1.0;
     for (int i = 0; i < phys_dmg_buffs.size(); ++i) {
-        float coefficient = 1.0 + float(phys_dmg_buffs[i]) / 100;
+        double coefficient = 1.0 + double(phys_dmg_buffs[i]) / 100;
         total_phys_dmg_mod = total_phys_dmg_mod * coefficient;
     }
 }
 
-float CharacterStats::get_mh_wpn_speed() {
+double CharacterStats::get_mh_wpn_speed() {
     return pchar->has_mainhand() ? equipment->get_mainhand()->get_base_weapon_speed() / haste_factor :
                                    2.0 / haste_factor;
 }
 
-float CharacterStats::get_oh_wpn_speed() {
+double CharacterStats::get_oh_wpn_speed() {
     return pchar->has_offhand() ? equipment->get_offhand()->get_base_weapon_speed() / haste_factor :
                                   300;
 }
