+++ Character/Class/General/OffhandAttack.h
@@ -7,8 +7,8 @@ class OffhandAttack: public Spell {
 public:
     OffhandAttack(Engine* engine, Character* pchar, CombatRoll* roll);
 
-    float get_next_expected_use() const;
-    void update_next_expected_use(const float);
+    double get_next_expected_use() const;
+    void update_next_expected_use(const double);
     bool attack_is_valid(const int) const;
     int get_next_iteration();
     void reset_effect() override;
@@ -19,9 +19,9 @@ public:
 protected:
     virtual void spell_effect() override;
     virtual void calculate_damage();
-    float next_expected_use;
+    double next_expected_use;
     int iteration;
-    QVector<float> talent_ranks;
+    QVector<double> talent_ranks;
 private:
 };
