+++ Character/Class/Hunter/Hunter.cpp
@@ -69,7 +69,7 @@ int Hunter::get_ap_per_agi() const {
     return 1;
 }
 
-float Hunter::global_cooldown() const {
+double Hunter::global_cooldown() const {
     return 1.5;
 }
