+++ Character/Class/Warlock/Warlock.cpp
@@ -68,7 +68,7 @@ int Warlock::get_ap_per_agi() const {
     return 1;
 }
 
-float Warlock::global_cooldown() const {
+double Warlock::global_cooldown() const {
     return 1.5;
 }
