+++ Character/Race/Races/Undead.cpp
@@ -29,11 +29,11 @@ int Undead::get_base_spirit(void) const {
     return 25;
 }
 
-float Undead::get_int_multiplier() const {
+double Undead::get_int_multiplier() const {
     return 1.0;
 }
 
-float Undead::get_spirit_multiplier() const {
+double Undead::get_spirit_multiplier() const {
     return 1.0;
 }
